pyproject.toml
[project]
name = "ledgerly"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "fastapi",
    "uvicorn[standard]",
    "jinja2",
    "python-multipart",
    "pdfplumber",
    "groq",
    "google-generativeai",
    "httpx",
    "supabase",
    "openpyxl",
    "apscheduler",
    "pydantic-settings",
    "resend",
    "python-Levenshtein",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-asyncio",
    "playwright",
    "ruff",
    "reportlab",
]

[tool.setuptools.packages.find]
include = ["app*"]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"
